Alteração de Controllers e Modelos
Adição de função de remoção de categoria no Controller/View de categorias.
Através da coluna selecionada, é possível filtrar resultados nos produtos (Ex.: Coluna Categoria, pesquisa por categoria).
Modelo de categoria inclui data de criação no construtor.
Criação de view sobre vendas.

+++ POO_TP_29559/Controllers/CategoriaController.cs
@@ -1,14 +1,18 @@
 ﻿using poo_tp_29559;
 using poo_tp_29559.Controllers;
 using poo_tp_29559.Models;
+using poo_tp_29559.Repositories;
 using poo_tp_29559.Views;
 using System.Collections.Generic;
 using System.Windows.Forms;
 
 public class CategoriaController : BaseController<Categoria, CategoriasForm>
 {
-    public CategoriaController(CategoriasForm view) : base(view, "Data/categorias.json")
+    private readonly ProdutoRepo _produtoRepo;
+
+    public CategoriaController(CategoriasForm view, ProdutoRepo produtoRepo) : base(view, "Data/categorias.json")
     {
+        _produtoRepo = produtoRepo;
         Initialize();
     }
 
@@ -22,10 +26,24 @@ public class CategoriaController : BaseController<Categoria, CategoriasForm>
         FiltrarItens(filtro, Categoria => Categoria.Nome != null && Categoria.Nome.ToLower().Contains(filtro.ToLower()));
     }
 
+    public void RemoveCategoria(Categoria item)
+    {
+        RemoveItem(item);
+    }
+
     protected override void RemoveItem(Categoria item)
     {
-        _repository.Remove(item);
-        CarregaItens();
+        var produtos = _produtoRepo.GetAll();
+
+        if (item.PodeSerEliminada(produtos))
+        {
+            _repository.Remove(item);
+            CarregaItens();
+        }
+        else
+        {
+            MessageBox.Show("Esta categoria não pode ser eliminada porque está associada a um ou mais produtos.");
+        }
     }
